requirements.txt
# HybridLogisticsHub - Dependencias
# Python 3.9+

# Framework API (FastAPI elegido por su rendimiento asíncrono)
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Driver PostgreSQL (psycopg2 como se requiere)
psycopg2-binary==2.9.9

# Driver MongoDB (PyMongo como se requiere)
pymongo==4.6.1

# Validación de datos
pydantic==2.5.2
pydantic-settings==2.1.0

# Utilidades
python-dotenv==1.0.0
